FTP: Fixed some bugs and did more test on ftp session
Former-commit-id: d8b7ac69f9b437fe46c90dfc9cff37958bd5454a
Former-commit-id: 8f503c74cec699f7cdafa6b3312d562695f9452f
Former-commit-id: 75f21ea4215436fde2b864df28cacf59b89a3f26

+++ src/mmt_tcpip/lib/protocols/ftp.h
@@ -56,7 +56,6 @@ static uint32_t ftp_connection_timeout = MMT_FTP_TIMEOUT * MMT_MICRO_IN_SEC;
 #define MMT_FTP_DATA_UPLOAD                 1 // The data connection is uploading data to server
 #define MMT_FTP_DATA_DOWNLOAD               2 // The data connection is downloading data from server
 
-
 /*
  * these are the commands we tracking and expecting to see to classify as FTP protocol
  */
@@ -178,7 +177,7 @@ typedef struct ftp_tuple6_struct{
  * FTP command structure: CMD PARAMETER
  */
 typedef struct ftp_command_struct{
-    int cmd;
+    uint16_t cmd;
     char *str_cmd;
     char *param;
 }ftp_command_t;
@@ -187,7 +186,7 @@ typedef struct ftp_command_struct{
  * FTP response structure
  */
 typedef struct ftp_response_struct{
-    int code;
+    uint16_t code;
     char *value;
 }ftp_response_t;
